refactored runtime.Debug -> util.Debug, so we can use Debug in runtime/ packages
