+++ commands/qemu-guest-tools/cmd.go
@@ -18,9 +18,10 @@ import (
 
 	"github.com/taskcluster/taskcluster-worker/commands"
 	"github.com/taskcluster/taskcluster-worker/runtime"
+	"github.com/taskcluster/taskcluster-worker/runtime/util"
 )
 
-var debug = runtime.Debug("guesttools")
+var debug = util.Debug("guesttools")
 
 func init() {
 	commands.Register("qemu-guest-tools", cmd{})
