+++ engines/mock/doc.go
@@ -2,6 +2,6 @@
 // but allows us to test plugins without having to run a real engine.
 package mockengine
 
-import "github.com/taskcluster/taskcluster-worker/runtime"
+import "github.com/taskcluster/taskcluster-worker/runtime/util"
 
-var debug = runtime.Debug("mockengine")
+var debug = util.Debug("mockengine")
