+++ engines/qemu/image/doc.go
@@ -3,6 +3,6 @@
 // the images don't reference external files as backing store.
 package image
 
-import "github.com/taskcluster/taskcluster-worker/runtime"
+import "github.com/taskcluster/taskcluster-worker/runtime/util"
 
-var debug = runtime.Debug("image")
+var debug = util.Debug("image")
