+++ engines/qemu/vm/doc.go
@@ -1,6 +1,6 @@
 // Package vm provides virtual machine abstractions for QEMU.
 package vm
 
-import "github.com/taskcluster/taskcluster-worker/runtime"
+import "github.com/taskcluster/taskcluster-worker/runtime/util"
 
-var debug = runtime.Debug("vm")
+var debug = util.Debug("vm")
