+++ runtime/util/debug.go
@@ -1,4 +1,4 @@
-package runtime
+package util
 
 import (
 	"fmt"
