+++ runtime/util/doc.go
@@ -1,0 +1,3 @@
+// Package util contains a few simple utilites that has no internal
+// dependencies. This is to encourage reuse of code.
+package util
